Fix issue with redirect
Previously, all redirects were done using HTTP 302.

Now they are done using HTTP 303 so that method would
reset to GET with no args.

This fixes a bug with deleting the pages, but still
the DELETE does not work quite right (does not follow
- or does not display - redirect to home page after
deletion, even though status code is OK and request
seems to be made)