Fix page delete
Give up on hopelessly bad browser support for DELETE
(they don't redirect properly back to homepage with
HTTP 303, and with 302 they repeat the DELETE on homepage).

So, we now do the delete action in the same way we do rename,
(Doing a POST, then checking for actual directive within the
parameters).