Add a working mediamap
Mediamap is an equivalent of sitemap, but reports
media files instead of pages.

All media files are attached to wiki pages of the
same name. The only difference is that the wiki
pages are in the pages/ subdirectory, and media files
are in the media/ subdirectory. From there the paths
are the same